chore: moved routes to fetch collection and environment into access-token module
