+++ packages/hoppscotch-backend/src/access-token/access-token.module.ts
@@ -2,9 +2,11 @@ import { Module } from '@nestjs/common';
 import { AccessTokenController } from './access-token.controller';
 import { PrismaModule } from 'src/prisma/prisma.module';
 import { AccessTokenService } from './access-token.service';
+import { TeamCollectionModule } from 'src/team-collection/team-collection.module';
+import { TeamEnvironmentsModule } from 'src/team-environments/team-environments.module';
 
 @Module({
-  imports: [PrismaModule],
+  imports: [PrismaModule, TeamCollectionModule, TeamEnvironmentsModule],
   controllers: [AccessTokenController],
   providers: [AccessTokenService],
   exports: [AccessTokenService],
